+++ src/cmd/extrinsics/instantiate.rs
@@ -17,9 +17,10 @@
 use super::{
     display_contract_exec_result, display_events, parse_balance, runtime_api::api,
     wait_for_success_and_handle_error, Balance, CodeHash, ContractAccount,
-    ContractMessageTranscoder, PairSigner, RuntimeApi, EXEC_RESULT_MAX_KEY_COL_WIDTH,
+    ContractMessageTranscoder, ExtrinsicOpts, PairSigner, RuntimeApi,
+    EXEC_RESULT_MAX_KEY_COL_WIDTH,
 };
-use crate::{name_value_println, util::decode_hex, ExtrinsicOpts, Verbosity};
+use crate::{name_value_println, util::decode_hex, Verbosity};
 use anyhow::{anyhow, Context, Result};
 use jsonrpsee::{
     types::{to_json_value, traits::Client as _},
@@ -63,10 +64,6 @@ pub struct InstantiateCommand {
     /// Maximum amount of gas to be used for this command
     #[structopt(name = "gas", long, default_value = "50000000000")]
     gas_limit: u64,
-    /// The maximum amount of balance that can be charged from the caller to pay for the storage
-    /// consumed.
-    #[structopt(long, parse(try_from_str = parse_balance))]
-    storage_deposit_limit: Option<Balance>,
     /// A salt used in the address derivation of the new contract. Use to create multiple instances
     /// of the same contract code from the same account.
     #[structopt(long, parse(try_from_str = parse_hex_bytes))]
@@ -129,7 +126,7 @@ impl InstantiateCommand {
         let args = InstantiateArgs {
             value: self.value,
             gas_limit: self.gas_limit,
-            storage_deposit_limit: self.storage_deposit_limit,
+            storage_deposit_limit: self.extrinsic_opts.storage_deposit_limit,
             data,
             salt,
         };
