+++ src/main.rs
@@ -40,8 +40,6 @@ use anyhow::{Error, Result};
 use colored::Colorize;
 use structopt::{clap, StructOpt};
 
-use sp_core::{crypto::Pair, sr25519};
-
 #[derive(Debug, StructOpt)]
 #[structopt(bin_name = "cargo")]
 #[structopt(version = env!("CARGO_CONTRACT_CLI_IMPL_VERSION"))]
@@ -69,45 +67,6 @@ impl std::str::FromStr for HexData {
 
     fn from_str(input: &str) -> std::result::Result<Self, Self::Err> {
         hex::decode(input).map(HexData)
-    }
-}
-
-/// Arguments required for creating and sending an extrinsic to a substrate node
-#[derive(Clone, Debug, StructOpt)]
-pub(crate) struct ExtrinsicOpts {
-    /// Path to the Cargo.toml of the contract
-    #[structopt(long, parse(from_os_str))]
-    manifest_path: Option<PathBuf>,
-    /// Websockets url of a substrate node
-    #[structopt(
-        name = "url",
-        long,
-        parse(try_from_str),
-        default_value = "ws://localhost:9944"
-    )]
-    url: url::Url,
-    /// Secret key URI for the account deploying the contract.
-    #[structopt(name = "suri", long, short)]
-    suri: String,
-    /// Password for the secret key
-    #[structopt(name = "password", long, short)]
-    password: Option<String>,
-    #[structopt(flatten)]
-    verbosity: VerbosityFlags,
-    /// Dry-run the extrinsic via rpc, instead of as an extrinsic. Chain state will not be mutated.
-    #[structopt(long, short = "rpc")]
-    dry_run: bool,
-}
-
-impl ExtrinsicOpts {
-    pub fn signer(&self) -> Result<sr25519::Pair> {
-        sr25519::Pair::from_string(&self.suri, self.password.as_ref().map(String::as_ref))
-            .map_err(|_| anyhow::anyhow!("Secret string error"))
-    }
-
-    /// Returns the verbosity
-    pub fn verbosity(&self) -> Result<Verbosity> {
-        TryFrom::try_from(&self.verbosity)
     }
 }
